Replace fmt.Sprintf in date/time formatting
These functions are very hot according to pprof,
with most program time being spent in them,
and their calls to Sprintf in particular.

This shaves off about 20% of runtime when parsing
the German GTFS dataset with gtfstidy.
In my opinion, the slightly more verbose code is
well worth this boost in performance.